+++ mne/connectivity/__init__.py
@@ -1,0 +1,5 @@
+""" Connectivity Analysis Tools
+"""
+
+from .utils import seed_target_indices
+from .spectral import spectral_connectivity